+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2020-04-17
+
 ### Added
 [#7] (https://github.com/AdrianInsua/neo4j-dashboard/issues/7) use d3 approximation and connect to neo4j-service.
 
@@ -19,5 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/AdrianInsua/neo4j-dashboard/compare/commits?targetBranch=0.1.0&sourceBranch=refs/heads/develop
+[Unreleased]: https://github.com/AdrianInsua/neo4j-dashboard/compare/commits?targetBranch=1.0.0&sourceBranch=refs/heads/develop
+[1.0.0]: https://github.com/AdrianInsua/neo4j-dashboard/compare/commits?targetBranch=0.1.0&sourceBranch=1.0.0
 [0.1.0]: https://github.com/AdrianInsua/neo4j-dashboard/compare/commits?until=0.1.0